feat(AEV-026): upgrade toolchain to GCC 15 / clang 21, split CI pipelines (#11)
* feat(AEV-026): split CI into PR and main pipelines, upgrade toolchain to GCC 15 / clang 21

- Replace ci.yml with pr.yml (format-check, tidy, build matrix, asan-test, build-windows)
  and main.yml (build matrix, asan-test only)
- PR pipeline checks out the synthetic merge commit to catch integration failures pre-merge
- clang-tidy runs in check-only mode (no --fix); format-check and tidy are PR-only
- Upgrade base image to ubuntu-26.04, GCC 15.2.0, clang-tidy-21, clang-format-21
- Update vcpkg builtin-baseline to 031f4afe03be385aee354f15be6e6b1fe57497c7
- Add missing release test preset to CMakePresets.json
- Add scripts/tidy.sh local runner mirroring the CI tidy job


* fix(AEV-026): fix all clang-tidy 21 warnings across source, tests, and bench

- Revert promise_type rename: auto-fix had renamed to PromiseType which
  broke coroutine machinery; restored with NOLINT(readability-identifier-naming)
- Wrap static test/bench helpers in anonymous namespaces to satisfy
  misc-use-anonymous-namespace across 15+ files
- Add NOLINT(bugprone-unchecked-optional-access) on Catch2 REQUIRE-guarded
  optional dereferences where tidy cannot see the guard
- Replace (void) discards with [[maybe_unused]] auto for [[nodiscard]]
  Asio returns (ex->run(), asio::write, asio::read)
- Add NOLINT(bugprone-unused-return-value) to s.connect() / s.close()
  calls where error_code is used for explicit error checking
- Change State enum base from int to uint8_t (performance-enum-size)
- Replace string concatenation with += in router test
  (performance-inefficient-string-concatenation)
- Use .at() for runtime-indexed array access in http-parse-roundtrip
  (cppcoreguidelines-pro-bounds-constant-array-index)
- Add NOLINT(bugprone-exception-escape) to bench and smoke main() functions
- Move misc-no-recursion NOLINT to correct line in router_impl.cpp

Build: clean (0 errors), 106/106 tests pass, 0 clang-tidy diagnostics.


* fix(AEV-026): fix remaining clang-tidy NOLINT placement issues

Use NOLINTNEXTLINE on the preceding line (not NOLINT on the same line)
for declarations that clang-format wraps across multiple lines, making
the suppression formatter-stable:
- app_impl.cpp: g_signal_executor global (cppcoreguidelines-avoid-non-const-global-variables)
- task.hpp: optional dereference guarded by exception_ check (bugprone-unchecked-optional-access)
- router_impl.cpp: recursive trie insert (misc-no-recursion)
- router.cpp: optional dereference after REQUIRE guard (bugprone-unchecked-optional-access)

Extract asio endpoint to a local variable so s.connect(ep, ec) fits
on one line, keeping the NOLINT(bugprone-unused-return-value) stable
across 7 test and bench files.


* Remove the NOLINT fixes with proper fixes

* feat(AEV-026): add clang 21 build matrix, self-hosted Linux runner, VS 2026

- Expand CI build matrix to 6 entries: GCC 15 x {debug,release} on Linux,
  clang 21 x {debug,release} on Linux, clang 21 (Homebrew llvm@21) x
  {debug,release} on macOS (Tier-2, continue-on-error)
- Add asan-test-clang job to both pr.yml and main.yml: ASan+UBSan under
  clang 21 on Linux as an independent sanitizer pass alongside GCC 15 ASan
- Add per-compiler vcpkg binary cache keys (gcc, clang-linux, clang-macos,
  msvc) to prevent ABI-incompatible binary reuse across compilers
- Fix configure steps to pass -DCMAKE_CXX_COMPILER="$CXX" explicitly,
  overriding the preset's hardcoded g++ cacheVariable for clang builds
- Switch all Linux jobs to self-hosted runner label self-ubuntu26.04
- Switch Windows build to windows-latest with Visual Studio 18 2026 generator
- Verified locally: clang 21.1.8 debug + ASan builds pass 106/106 tests


* fix(AEV-026): replace Kitware PPA cmake install with system apt

apt-key is removed in Ubuntu 26.04 and the PPA codename was wrong
(plucky is 25.04; 26.04 is resolute). Ubuntu 26.04 ships cmake 3.31+
which exceeds the project minimum of 3.27, so the PPA is not needed.

Replaced all cmake install blocks in pr.yml and main.yml with a plain
`sudo apt-get install -y cmake` from the system repository.


* fix(AEV-026): fix MSVC generator name and macOS deployment target

MSVC: windows-latest ships VS 2022; revert CMakePresets.json generator
from the non-existent "Visual Studio 18 2026" back to "Visual Studio 17 2022".

macOS: std::move_only_function in libc++ is gated by the macOS deployment
target. Set MACOSX_DEPLOYMENT_TARGET=14.0 in GITHUB_ENV from the Homebrew
install step so CMake picks it up at configure time. macOS 14 is the
minimum version where libc++ exposes all required C++23 features used by
the codebase (move_only_function, etc.).


* Fix the signal in MSVC

* remove mac os build from plans

* remove document section from CLAUDE.md since it does not exist

---------

+++ src/http/http_parser.hpp
@@ -41,7 +41,7 @@ namespace aevox::detail {
 constexpr std::size_t kParserDefaultMaxHeaderCount{100};
 // Default maximum body size for the parser's safety net (1 MiB).
 // Overridden per-connection from AppConfig::max_body_size.
-constexpr std::size_t kParserDefaultMaxBodyBytes{1u * 1024u * 1024u};
+constexpr std::size_t kParserDefaultMaxBodyBytes{static_cast<std::size_t>(1u * 1024u * 1024u)};
 
 // =============================================================================
 // ParseError
